{{cookiecutter.project_name}}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.project_name }}"
description = "{{ cookiecutter.project_verbose_name }}"
version = "0.1.0"
readme = "README.md"
authors = ["{{ cookiecutter.project_domain }}"]


[tool.poetry.dependencies]
python = "3.9.13"

django = { version = "^3.2", extras = ["argon2"] }
django-split-settings = "^1.1"
django-axes = "^5.39"
django-csp = "^3.7"
django-health-check = "^3.16"
django-http-referrer-policy = "^1.1"
django-permissions-policy = "^4.13"
django-stubs-ext = "^0.5"

psycopg2-binary = "^2.9"
gunicorn = "^20.0"
python-decouple = "^3.6"
structlog = "^22.1"

# Docs extra:
sphinx = { version = "^5.1", optional = true }
sphinx-autodoc-typehints = { version = "^1.19", optional = true }
tomli = { version = "^2.0", optional = true }
sentry-sdk = "^1.9.8"
celery = "^5.2.7"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.6"
django-querycount = "^0.7"
django-migration-linter = "^4.1"
django-extra-checks = "^0.12"
django-coverage-plugin = "^2.0"
nplusone = "^1.0"

wemake-python-styleguide = "^0.16"
flake8-pytest-style = "^1.6"
flake8-django = "^1.1"
flake8-logging-format = "^0.7"
nitpick = "^0.32"
doc8 = "^1.0"

pytest = "^7.1"
pytest-django = "^4.5"
pytest-cov = "^3.0"
pytest-randomly = "^3.12"
pytest-deadfixtures = "^2.2"
pytest-timeout = "^2.1"
django-test-migrations = "^1.2"
hypothesis = "^6.54"

mypy = "^0.971"
django-stubs = "^1.12"

yamllint = "^1.27"
safety = "^2.1"
dotenv-linter = "^0.3"
polint = "^0.4"
dennis = "^1.1"
dump-env = "^1.3"
ipython = "^8.4"
celery = {extras = ["redis"], version = "^5.2.7"}

[tool.poetry.extras]
docs = ["sphinx", "sphinx-autodoc-typehints", "tomli"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"
